imp(evm): update move gas functions to a new file
